rt - many changes
